HACK: include themwi-rtp-lib inside osmo-bts
The author's eventual goal is to get themwi-rtp-lib accepted into
Osmocom proper, perhaps to be included in libosmo-netif.  However,
before that idea can be formally proposed, much less implemented,
the new RTP endpoint library needs to be tested.  One of the key
tests to be performed is suitability of this new library to serve
as a replacement for libortp in osmo-bts - hence a test version
of osmo-bts needs to be built with themwi-rtp-lib prior to any
proposed libosmo-netif integration.

To ease testing on sysmoBTS hardware, which entails building test
versions of osmo-bts-sysmo with sysmoBTS SDK, the entire themwi-rtp-lib
code is imported inside osmo-bts source tree, as opposed to
implementing a dependency mechanism whereby Autotools-based
configure and makefiles would pick up an external non-Osmocom
library.

The version of themwi-rtp-lib embedded here corresponds to
Hg commit 36:84affc6de365.

Change-Id: I1eea83351ad61236d1b2b40ad3d0f66660ed79de