HACK: include local copy of twrtp
The version of twjit & twrtp included here exactly matches
Patchset 5 in these Gerrit patches:

https://gerrit.osmocom.org/c/libosmo-netif/+/39280
https://gerrit.osmocom.org/c/libosmo-netif/+/39281

I am embedding a local copy of this twrtp into a developer branch
of osmo-bts because I want to be able to use twrtp on sysmoBTS.
With sysmoBTS SDK one can easily build a custom version of
osmo-bts-sysmo that still uses only unmodified mainline libosmo*
libraries, but building an application that requires a branch
version of a library is much harder.

All names that begin with osmo_twjit or osmo_twrtp in the
libosmo-netif merge candidate version have been renamed to
obts_twjit or obts_twrtp in this embedded version, so there is
no name collision.

Change-Id: I4ee481f9f9122f3db3d7bddc26e19e84b77b80ef